requirements.txt
# Core dependencies
jupyter>=1.0.0
notebook>=6.4.0
ipywidgets>=7.6.0

# AWS SDK
boto3>=1.26.0
botocore>=1.29.0

# LangGraph and LangChain
langgraph>=0.0.40
langchain>=0.1.0
langchain-aws>=0.1.0

# Data processing
pandas>=1.5.0
numpy>=1.21.0
pillow>=9.0.0

# Visualization
plotly>=5.0.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Utilities
python-dotenv>=0.19.0
uuid>=1.30
base64>=1.0.0
json5>=0.9.0

# Optional: Email notifications
email-validator>=1.3.0

# Development
ipykernel>=6.0.0